backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["jgabriel1 <josegabriel@det.ufc.br>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.55.1"
uvicorn = "^0.11.5"
pymongo = {extras = ["tls"], version = "^3.10.1"}
dnspython = "^1.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.2"
python-dotenv = "^0.13.0"
flake8 = "^3.8.2"

[tool.poetry.scripts]
start = "scripts:start"
dev = "scripts:dev"
test = "scripts:test"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
